// src/store/types/TermFullJourney/TDropdownData.ts
import { TDropdown } from "./TDropdown";

export interface TDropdownData {
  state: TDropdown[];
  MAX_ANNUAL_INCOME: TDropdown[];
  MAX_SUM_ASSURED: TDropdown[];
  PREGNANCY_WEEKS: TDropdown[];
  purposeTravel: TDropdown[];
  typeResidence: TDropdown[];
  modeTravel: TDropdown[];
  AADHAR_OPTIONS_DATA: TDropdown[];
  IT_PROOF_MASTER: TDropdown[];
  IDENTITY_PROOF_MASTER: TDropdown[];
  EDUCATION_MASTER: TDropdown[];
  BENEFICIARY_RELATIONSHIP: TDropdown[];
  TRUSTEE_TYPE: TDropdown[];
  buyingPolicy: TDropdown[];
  eia: TDropdown[];
  vaccine: TDropdown[];
  myProf: TDropdown[];
  indsType: TDropdown[];
  orgType: TDropdown[];
  occupationData: TDropdown[];
  genderData: TDropdown[];
  maritalStatusData: TDropdown[];
  nomineeRelationshipData: TDropdown[];
  annualIncomeData: TDropdown[];
  heightFeetData: TDropdown[];
  heightInchesData: TDropdown[];
  tobacco: TDropdown[];
  alcohol: TDropdown[];
  addressProf: TDropdown[];
  ageProf: TDropdown[];
  incomeProf: TDropdown[];
  identityProf: TDropdown[];
  income_tax_proof: TDropdown[];
  insuranceComp: TDropdown[];
  country: TDropdown[];
  HDFC_CKYC_OCCUPATION: TDropdown[];
  insurance_company: TDropdown[];
  HDFC_MEDICAL_DESIGNATION: TDropdown[];
  NATURE_OF_WORK: TDropdown[];
  COMPANY_BELONGS: TDropdown[];
  SELECT_ANY_SIMULTANEOUS: TDropdown[];
  HDFC_COUNTRIES: TDropdown[];
  HOBBIES: TDropdown[];
  MAY_UNDERGO_MEDICAL_EXAMINATION: TDropdown[];
  CONSUME_ALCOHOL: TDropdown[];
  TOBACCO_PRODUCT: TDropdown[];
  RELATION_OF_LIFE_TO_BE_ASSURED: TDropdown[];
  HDFC_DISEASE: TDropdown[];
  AGE_AT_DIAGNOSIS: TDropdown[];
  HDFC_ALIVE: TDropdown[];
  HDFC_CURRENT_AGE: TDropdown[];
  HDFC_HEIGHT: TDropdown[];
  HDFC_WEIGHT: TDropdown[];
  have_you_suffered_from_one: TDropdown[];
  have_you_suffered_from_two: TDropdown[];
  have_you_suffered_from_three: TDropdown[];
  have_you_suffered_from_four: TDropdown[];
  have_you_suffered_from_five: TDropdown[];
}